unifarma_shipping_system/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.3.6
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.19
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-timezone-field==7.1
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
kombu==5.4.2
packaging==24.2
prompt_toolkit==3.0.50
psycopg2-binary==2.9.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
PyYAML==6.0.2
redis==5.0.1
requests==2.31.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
